Add histogram function to ThreadSafeCache
A new function `histogram` was added to `ThreadSafeCache`. This function allows for the visual representation of the distribution of elements in the cache, which helps in debugging and performance tracking. It prints a histogram where each bucket is represented by a line with a number of stars corresponding to the number of elements in the bucket.