Optimize type casting and container operations
Removed unnecessary union for casting types, replaces it with memcpy for safety reasons. Adjusted `emplace_front` and `find_or_create_impl` calls to use `std::move` instead of `std::forward` for clarity. The decision to use `std::move` is to express clearly that we are transferring ownership, and we expect the key/values not to be used afterward which should generally be the case in insert or update operations.