Refactor and optimize map iterators and hashing
Streamlined comments and formatting for better readability within `ConcurrentUnorderedMap.h`. Introduced an `m_end_iterator` boolean to more cleanly handle end-of-iteration cases. The value of `m_end_iterator` is set upon initialization and updated when the iterator has traversed all elements. This improvement eradicates any need for acquiring and holding a lock when checking for end-of-iteration, thus enhancing performance.

Also addressed potential issues with std::hash's identity function causing small hash values by multiplying hash values with a large co-prime number. This prevents hashing collision and results in a better spread in the map.

Replaced value-initialized iterators with an iterator created using the map itself in the methods: begin() and end(). Ensuring the iterators used are associated with the map.